// week9/patientiorBE/types.ts
export interface Patient {
    id: string;
    name: string;
    dateOfBirth: string;
    gender: string;
    occupation: string;
    ssn?: string,
    entries?: Entry[]
}

export interface NewPatient {
    name: string;
    dateOfBirth: string;
    gender: string;
    occupation: string;
    ssn?: string,
    entries: Entry[]
}

export type NonSensitivePatient = Omit<Patient, 'ssn' | 'entries'>


export enum Gender {
    Male = "male",
    Female = "female",
    Other = "other"
}

export interface NewHospitalEntry {
    type: "Hospital";
    discharge?: {
        date: string,
        criteria: string
    }
    description: string;
    date: string,
    specialist: string;
}

interface BaseEntry {
    id: string;
    description: string;
    date: string;
    specialist: string;
    diagnosisCodes?: Array<Diagnosis['code']>;
}

export interface Diagnosis {
    code: string;
    name: string;
    latin?: string;
}

export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3
}

export interface HealthCheckEntry extends BaseEntry {
    type: "HealthCheck";
    healthCheckRating: HealthCheckRating;
}

export interface HospitalEntry extends BaseEntry {
    type: "Hospital";
    discharge?: {
        date: string,
        criteria: string
    }
}

interface OccupationalHealthcareEntry extends BaseEntry {
    type: "OccupationalHealthcare";
    employerName: string;
    sickLeave?: {
        startDate: string,
        endDate: string
    }
}

export type Entry = | HospitalEntry | OccupationalHealthcareEntry | HealthCheckEntry;